requirements.txt
# 精准LBS社交匹配系统依赖包

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
redis==5.0.1
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# 机器学习
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# 地理计算
geopy==2.4.0
shapely==2.0.2

# 工具库
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 监控和日志
prometheus-client==0.0.19
structlog==23.2.0
psutil==5.9.5

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
